+++ packages/react-templates/src/components/FeedObjectDetail/FeedObjectDetail.stories.tsx
@@ -14,7 +14,7 @@ const template = (args) => (
   </div>);
 
 const templateContainerFixed = (args) => (
-	<div style={{position: 'fixed', bottom: 0, left: 0, right: 0, zIndex: 1000, maxWidth: '1200px', height: '92vh', overflow: 'auto', paddingLeft: 20, paddingRight: 20}} id="scrollableDiv">
+	<div style={{position: 'fixed', bottom: 0, left: 0, right: 0, zIndex: 1, maxWidth: '1200px', height: '92vh', overflow: 'auto', paddingLeft: 20, paddingRight: 20}} id="scrollableDiv">
 		<FeedObjectDetailTemplate {...args} />
 	</div>);